"RISC-V_Generic/libdl: Do another relocation parse for LO12 relocations
LO12 relocations point to HI20 relocation labels that hold the actual
symvalue to be relocated to. They don't need to be sequential. So, the first
relocation parses HI20 ones, adds them to the list and ignores LO12 ones. Then,
another parse just for LO12 relocations is performed, and the added HI20
relocation details are used.

Also, the local LO12 symbols are extracted from the local table on the go as
a workaround to handle the case when there are duplicate local labels generated
by Clang and lld -r to create a single object module compartment from multiple
source files.