libvirtio: virtio-net - wait until the packet is sent by the device
The driver doesn't have TX buffers and rather uses whatever buffer is passed to
it from the TCP/IP stack which might be freed/used on return. For this reason,
wait until the device has sent/processed the packet before returning from this
function.

This is an inefficient implementation and increases latency. A proper solution
would check the packet is properly sent on the next call.