FreeRTOS-libdl: Add support for FreeRTOS-FAT file system
This adds support to load and open objects from a file system rather than from
the ELF image.

file handlers in FreeRTOS-FAT are pointers, so change fds to be void pointers
while casting them to FF_FILE when ipconfigUSE_FAT_LIBDL is set, otherwise
preserve backward compatibility assuming fds refer to compartment numbers
and a compartment/object is embedded in the ELF image.