RISC-V-Generic: Optionally dynamically load libraries as compartments
* Add a compartmentalize switch
* Compartmentalication requires libdl (dynamic loader) and libfat (filesystem)
* During build tume, the libraries that need to be compartmentalized are
defined by the wscript, then, they get built and embedded in the img filesystem
by converting each lib to a hex C array and writing it to the filesystem.
* main does have a dynamic loader (libdl) that dlopen the main obj file from
the app, which searches for all object and lib dependencies, loads them, and
fix their relocations.
* Loaded libraries can be updated in the filesystem and if that happens, they
will be re-loaded on the next dlopen call.
* Currently, as a PoC, main_servers dynamically loads:
  main_servers (app), virtio, tcp/ip stack, and cli as libs.