link.ld: Remove custom prog headers
Custom prog headers were added to be able to embed and load the symbol table
by loaders that is needed for libdl during dynamic loading and linking. This
works on some targets that blindly load symtables indicated by prog headers,
but not others (such as GDB) that drop loading (in target memory) any sections
having SYMTAB or STRTAB attributes. This is now fixed in 19423f2 by separately
extracting the symbol table and do a 2-pass linking to embed the symbol table
in rodata the same way as RTEMS does, using rtems-tools.