wscript: Don't assume Python 3
Waf itself runs with whatever python is. Traditionally this was python2,
and for many systems that is still the case, though some systems (and
presumably the ones on which this has been used so far) now have it
point to python3, for better or worse. Waf does not execute wscript, it
sources it, so the shebang here is unused and misleading. Thus, delete
it and stick to valid Python 2 syntax.