wscript: Fix bin2header for Python 2
For Python 3, reading the file (and encoding a string) gives a byte
array so each element is an actual integer and thus chr(ord(...)) is a
complete waste of time. For Python 2 that doesn't work because there are
no byte strings, everything is a string and you just have to know
whether it's encoded or decoded, so elements of byte strings are still
strings, not integers, and need unpacking.