wscript: Fix build on macOS
The ARCH variable is pre-defined by waf, and is used specifically on
macOS for the -arch command-line option to Clang (which is akin to
-target and various sysroot options, allowing developers with a macOS
SDK, or iOS, to compile their code for a specific supported
architecture). It's interpreted as an array, so that you can build
universal binaries easily (with fat object files), meaning that using
the value "riscv64" gets interpreted as architectures r, i, s, and so
on, resulting in a spew of warnings that these are being ignored.