wscript: Ensure system compiler is used for libdl-tools build
Using the CHERI-LLVM cross-compiler does not work on macOS/arm64, and
possibly not in macOS/x86_64 either. By default waf will look for clang
in PATH on macOS. Since the freertos cheribuild target bogusly puts the
SDK directory in PATH rather than providing absolute paths to everything
like we normally to (to avoid issues like this) we need to ensure the
SDK dir's clang is not found. To do this, we exploit the fact that the
SDK directory has clang and $triple-clang symlinks but no cc symlink,
and thus CC=cc lets us pick up the system compiler.