wscript[s]: Set code model to medlow to force relocs on all function pointers
Courtsey to jrtc27 for guidance on this. In order to emit domain switch trampolines for all function
pointers, the dynamic loader needs to detect those by a relocation. With mcmodel=medium, it does
only emit relocations on pointers to global functions, but not static/local ones. medlow emits
absolute HI20/LO12 relocs for all function pointers regardless of they are static or not, which is
what the dynamic linker requires.