Add IMPLICIT_MEM_0 option to disable unnecessary bss, sbss zeroing
Many RISC-V simulators start with memory implicitly zeroed. In these cases
it can be unnecessary and extremely slow to zero out the .bss and .sbss
sections. This commit adds --implicit-mem-0 to the wscript options,
which defines IMPLICIT_MEM_0 in boot.S, which disables the 'fill_block'
loops for those sections.

Stack initialization with 0xABABABAB is not disabled under IMPLICIT_MEM_0,
because it isn't zero-initialization.